slab: cast sizeof() result to int.

+++ tests/test_slab.c
@@ -61,8 +61,8 @@ int main(void)
 {
 	slab_init(&slb, sizeof(struct item));
 
-	printf("objsize     = %d (original %ld)\n",
-		slb.objsize, sizeof(struct item));
+	printf("objsize     = %d (original %d)\n",
+		slb.objsize, (int)sizeof(struct item));
 	printf("count       = %d\n", slb.count);
 	printf("info_offset = %d\n", slb.info_offset);
 	printf("slabsize    = %d\n", slb.slabsize);
